refactor: Simplify _id to id resolving
