+++ src/common/common.graphql
