+++ src/common/id-resolver.ts
@@ -1,0 +1,8 @@
+import { Parent, ResolveField } from '@nestjs/graphql';
+
+export class IdResolver {
+  @ResolveField('id')
+  getId(@Parent() { _id }) {
+    return _id;
+  }
+}
